[parser] Move parser directory to lib/parser
Preferred since lints happen in lib directory.
