+++ lib/parser/util/Environment.js
@@ -11,9 +11,9 @@ const _size = new WeakMap();
 export class Environment {
   /**
    * Constructs a new persistent linked list environment.
-   * @param {*} key 
-   * @param {*} value 
-   * @param {*} tail 
+   * @param {*} key
+   * @param {*} value
+   * @param {*} tail
    * @throws {Error}
    */
   constructor (key, value, tail) {
