+++ lib/parser/util/LinkedList.js
